moving js to dedicated file
