+++ js/tallstackui.js
@@ -9,6 +9,7 @@ import modal from './components/modal/modal';
 import number from './components/form/number';
 import select from './components/select/select';
 import slide from './components/slide/slide';
+import pin from './components/form/pin';
 import tab from './components/tab/tab';
 import textArea from './components/form/textarea';
 import toastBase from './components/toast/toast-base';
@@ -19,12 +20,13 @@ document.addEventListener('alpine:init', () => {
   window.Alpine.plugin(tooltip);
   window.Alpine.data('tallstackui_banner', banner);
   window.Alpine.data('tallstackui_formColor', color);
+  window.Alpine.data('tallstackui_formNumber', number);
+  window.Alpine.data('tallstackui_formPin', pin);
   window.Alpine.data('tallstackui_darkTheme', darkTheme);
   window.Alpine.data('tallstackui_dialog', dialog);
   window.Alpine.data('tallstackui_dropdown', dropdown);
   window.Alpine.data('tallstackui_loading', loading);
   window.Alpine.data('tallstackui_modal', modal);
-  window.Alpine.data('tallstackui_formNumber', number);
   window.Alpine.data('tallstackui_select', select);
   window.Alpine.data('tallstackui_slide', slide);
   window.Alpine.data('tallstackui_tab', tab);
